requirements.txt
opencv-python==4.1.2.30
numpy==1.17.2
Keras==2.3.1
Flask==1.1.1
tensorflow==1.14.0
requests==2.22.0
matplotlib==3.1.1
Pillow==7.0.0
click==7.0